% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{amphibiansBR}
\alias{amphibiansBR}
\title{Brazilian Amphibian Species List Dataset}
\format{
\subsection{\code{amphibiansBR}}{

A data frame with 1,188 rows and 8 columns:
\describe{
\item{order}{Taxonomic Order}
\item{family}{Taxonomic Family}
\item{subfamily}{Taxonomic Family}
\item{sp_num}{Species number on list}
\item{genus}{Genus}
\item{epithet}{Specifc epithet}
\item{author}{Authorship of original description}
\item{status}{General remarks about species}
...
}
}
}
\source{
\url{https://zenodo.org/record/4716176}
}
\usage{
amphibiansBR
}
\description{
A dataset of the official Brazilian Amphibian Species List publish
by the Brazilian Herpetolgical Society (SBH)
}
\keyword{datasets}
