% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Segalla_2021}
\alias{Segalla_2021}
\title{Brazilian Amphibian Species List Dataset}
\format{
\subsection{\code{Segalla_2021}}{

A data frame with 1,188 rows and 8 columns:
\describe{
\item{species_id}{Id of species in list}
\item{order}{Taxonomic Order}
\item{family}{Taxonomic Family}
\item{subfamily}{Taxonomic Family}
\item{genus}{Genus}
\item{epithet}{Specifc epithet}
\item{species}{Scientific name}
\item{author}{Authorship of original description}

...
}
}
}
\source{
\url{https://zenodo.org/record/4716176}
}
\usage{
Segalla_2021
}
\description{
A dataset of the official Brazilian Amphibian Species List publish
by the Brazilian Herpetological Society (SBH).
}
\keyword{datasets}
